优化API的请求管理
